requirements.txt
fastapi
uvicorn[standard]
pdf2image
pillow
python-multipart
jinja2
